refactored some part and wrote test codes

+++ api/databases.py
@@ -65,17 +65,23 @@ class UserDatabaseOperation(User, DatabaseOperation):
     
     def get_user(self, usrname=""):
         """get all users from the database"""
-        if usrname == "":
-            return self.init_user_db.filter_by(username=User.username).all()
+        if not isinstance(usrname, str):
+            raise ValueError("Username must be a string")
         else:
-            return self.init_user_db.filter_by(username=User.username). \
-                filter(usrname == User.username).first()
+            if usrname == "":
+                return self.init_user_db.filter_by(username=User.username).all()
+            else:
+                return self.init_user_db.filter_by(username=User.username). \
+                    filter(usrname == User.username).first()
     
     def delete_user(self, username):
         """delete a user in the database"""
-        if username:
-            user = self.get_user(usrname=username)
-            self.init_db.removeFromDbSession(user)
+        if not isinstance(username, str):
+            raise ValueError("Username must be a string")
+        else:
+            if username:
+                user = self.get_user(usrname=username)
+                self.init_db.removeFromDbSession(user)
             
     def add_user(self, user):
         if user:
@@ -90,11 +96,14 @@ class TaskDatabaseOperation(DatabaseOperation, Task):
     
     def get_task(self, task_title=""):
         "fetch user tasks from the db"
-        if task_title == "":
-            return self.init_task_db.filter_by(title=Task.title).all()
+        if not isinstance(task_title, str):
+            raise ValueError("Task title must be a string")
         else:
-            return self.init_task_db.filter_by(title=task_title).\
-                filter(Task.title == task_title).first()
+            if task_title == "":
+                return self.init_task_db.filter_by(title=Task.title).all()
+            else:
+                return self.init_task_db.filter_by(title=task_title).\
+                    filter(Task.title == task_title).first()
     
     def add_new_task(self, task):
         """add new task to the database"""
@@ -103,17 +112,23 @@ class TaskDatabaseOperation(DatabaseOperation, Task):
             
     def delete_task(self, task_title=""):
         """delete a task with task id"""
-        if task_title:
-            task = self.get_task(task_title=task_title)
-            self.init_db.removeFromDbSession(task)
+        if not isinstance(task_title, str):
+            raise ValueError("Task title must be a string")
+        else:
+            if task_title:
+                task = self.get_task(task_title=task_title)
+                self.init_db.removeFromDbSession(task)
         
-    def edit_task(self, task_title, **kwargs):
+    def edit_task(self, task_title="", **kwargs):
         """edit task with task id"""
-        if task_title:
-            task = self.get_task(task_title=task_title)
-            task.title = kwargs.get('title')
-            task.description = kwargs.get('description')
-            self.task_session.commit()
+        if not isinstance(task_title, str) or kwargs:
+            raise ValueError("function arguments must be <str> <dict>")
+        else:
+            if task_title:
+                task = self.get_task(task_title=task_title)
+                task.title = kwargs.get('title')
+                task.description = kwargs.get('description')
+                self.task_session.commit()
 
 class TeamMemberDatabaseOperation(DatabaseOperation, TeamMember):
     """perform CRUD operation on the database for team member model"""
@@ -123,13 +138,16 @@ class TeamMemberDatabaseOperation(DatabaseOperation, TeamMember):
     
     def get_team_member(self, user_id=None, task_id=None, team_name="", task_name=""):
         """fetche team members in the database"""
-        if team_name == "" and task_name == "":
-            return self.init_team_db.filter_by(team_name=TeamMember.team_name).all()
+        if not isinstance((team_name, task_name), (str, str)):
+            raise ValueError("arguments must strictly be strings")
         else:
-            return self.init_team_db.filter_by(team_name=TeamMember.team_name).\
-                                                filter(TeamMember.task_id == task_id and
-                                                       TeamMember.user_id == user_id and 
-                                                       TeamMember.task_name == task_name).first()
+            if team_name == "" and task_name == "":
+                return self.init_team_db.filter_by(team_name=TeamMember.team_name).all()
+            else:
+                return self.init_team_db.filter_by(team_name=TeamMember.team_name).\
+                                                    filter(TeamMember.task_id == task_id and
+                                                        TeamMember.user_id == user_id and 
+                                                        TeamMember.task_name == task_name).first()
         
     def add_new_team_member(self, team_member):
         """add a new team member"""
@@ -141,4 +159,3 @@ class TeamMemberDatabaseOperation(DatabaseOperation, TeamMember):
         if team_member:
             self.init_db.removeFromDbSession(team_member)
     
-
